+++ css/CssGrid/example.html
@@ -7,7 +7,7 @@
 </head>
 <body>
     <script rel="preload" type="module">
-        import CssGrid from '../CssGrid.js'
+        import CssGrid from './CssGrid.js'
         customElements.define('css-grid', CssGrid)
     </script>
     <css-grid>
